+++ src/components/Nav.jsx
@@ -1,16 +1,25 @@
-import React from 'react'
+import React, { useState } from 'react'
 import * as S from "./Nav_styled";
 import {BiSolidMoon} from "react-icons/bi"
 import {BsFillSunFill} from "react-icons/bs"
+import Login from './Login/Login';
 
-function Nav({ openModal }) {
- 
+function Nav() {
+  const [modal, setModal] = useState(false);
+  
+  const openModal = () => {
+    setModal(true);
+  };
+
+  const closeModal = () => {
+    setModal(false);
+  };
 
   return (
     <>
-       <S.NavBox>
+      <S.NavBox>
           <h4>항구</h4>
-          <div  style={{ display: 'flex', alignItems: 'center' }}>
+          <div style={{ display: 'flex', alignItems: 'center' }}>
             <div>
               <BsFillSunFill size="27" />
               <BiSolidMoon size="27" />
@@ -18,8 +27,11 @@ function Nav({ openModal }) {
             <S.StBtn onClick={openModal}>로그인</S.StBtn>
           </div>
       </S.NavBox>
-  
-   </>
+      <Login
+        modal={modal}
+        closeModal={closeModal}
+      />
+    </>
   )
 }
